--- a/src/lib.rs
+++ b/src/lib.rs
@@ -41,8 +41,6 @@
 #![doc(html_logo_url = "https://unicode-rs.github.io/unicode-rs_sm.png",
        html_favicon_url = "https://unicode-rs.github.io/unicode-rs_sm.png")]
 
-extern crate smallvec;
-
 pub use tables::UNICODE_VERSION;
 pub use decompose::Decompositions;
 pub use quick_check::{