@Manishearth 2019-09-19
Yeah SmallVec is used specifically for its perf benefit here. Note that none of those CVEs affect this crate, which uses relatively simple APIs.

We should audit the crate further and try to run it with `miri`, but we aren't going to remove it.

Folks may be okay with a feature flag.