@Shnatsel 2020-03-16
For posterity, #54 has replaced use of SmallVec with safe code without regressing performance.